src/tts/requirements.txt
torch==1.13.0
torchaudio==0.13.0
huggingface-hub==0.10.1
fairseq==0.12.2
redis==4.3.4
python-dotenv==0.21.0
phonemizer==3.2.1
tensorboardX==2.5.1
sentencepiece==0.1.97